refactor: core 代码整理——去重共享工具、收敛类型、memory/render 归位、PathRule 移至 access
