+++ packages/core/src/capabilities/builtin.ts
@@ -1,7 +1,8 @@
 import type { Capability } from "../kernel/capability.js";
 import { fsCapability } from "./fs/index.js";
 import { skillCapability } from "./skill/index.js";
-import { changelogCapability, renderCapability } from "./changelog/index.js";
+import { changelogCapability } from "./changelog/index.js";
+import { renderCapability } from "./render/index.js";
 import { agentMgmtCapability } from "./agent-mgmt/index.js";
 import { interactionCapability } from "./interaction/index.js";
