+++ packages/core/src/kernel/ports.ts
@@ -2,6 +2,8 @@ import type { ApprovalGate, AskGate } from "./gates.js";
 import type { Logger } from "../logger.js";
 import type { FileWriteMutex } from "../utils/file-write-mutex.js";
 import type { ProjectStore } from "../store/project.js";
+import type { AgentProfile } from "../types.js";
+import type { PathRule } from "../access/path-category.js";
 
 export type SessionEventPayload =
   | { type: string; [key: string]: unknown }
@@ -17,12 +19,6 @@ export interface SessionPort {
   ): Promise<void>;
   abortSession(sessionId: string): void;
   sessionExists(agentId: string, sessionId: string): boolean;
-}
-
-export interface PathRule {
-  readonly match: RegExp;
-  readonly category: string;
-  readonly llm: { readonly read: boolean; readonly write: boolean };
 }
 
 export interface AgentStoreScope {
@@ -55,7 +51,7 @@ export interface ToolCatalog {
 export interface ToolHost {
   readonly agentId: string;
   readonly sessionId: string;
-  readonly profile: import("../types.js").AgentProfile;
+  readonly profile: AgentProfile;
   readonly projectRoot: string;
   readonly projectStore: ProjectStore;
   readonly fileWriteMutex: FileWriteMutex;
@@ -69,7 +65,7 @@ export interface ToolHost {
 
 export interface SessionView {
   readonly agentId: string;
-  readonly profile: import("../types.js").AgentProfile;
+  readonly profile: AgentProfile;
   readonly projectStore: ProjectStore;
   readonly stores: StoreRegistry;
 }
