+++ packages/core/src/session/agent-assembly.ts
@@ -10,6 +10,7 @@ import { readContextFiles, type ContextFile } from "../context/read-context-file
 import type { ToolHost } from "../kernel/ports.js";
 import { serializeBlocks, type ContextBlock } from "../kernel/context-block.js";
 import { llmPolicyOf } from "../capabilities/shared/llm-policy.js";
+import { escapeXmlAttr } from "../utils/xml-escape.js";
 import type { RuntimeDeps } from "./runtime.js";
 
 export function composeStreamFn(
@@ -63,14 +64,6 @@ interface SessionMeta {
   sessionId: string;
 }
 
-function escapeAttr(value: string): string {
-  return value
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;");
-}
-
 export function buildProjectInstructions(content: string): ContextBlock | null {
   if (content.trim() === "") return null;
   return {
@@ -108,7 +101,7 @@ export function buildPreloadedContext(files: ContextFile[]): ContextBlock | null
     kind: "preloaded-context",
     render: () => {
       const rendered = files
-        .map((f) => `<context-file path="${escapeAttr(f.path)}">\n${f.content}\n</context-file>`)
+        .map((f) => `<context-file path="${escapeXmlAttr(f.path)}">\n${f.content}\n</context-file>`)
         .join("\n");
       return `<preloaded-context>\n${rendered}\n</preloaded-context>`;
     },
