+++ packages/core/src/tools/copy-file.ts
@@ -3,11 +3,9 @@ import fsSync from "node:fs";
 import path from "node:path";
 import { Type } from "@sinclair/typebox";
 import type { AgentTool } from "@earendil-works/pi-agent-core";
-import type { AccessPolicy } from "../access/access-policy.js";
+import type { AccessPolicyProvider } from "../access/access-policy.js";
 import type { FileWriteMutex } from "../utils/file-write-mutex.js";
 import { resolveProjectPath } from "../utils/path-safety.js";
-
-type AccessPolicyProvider = () => AccessPolicy;
 
 const CopyFileParams = Type.Object({
   source: Type.String({ description: "Source file path relative to project root" }),
